# docs/swagger.yaml
basePath: /api/v1
definitions:
  internal_api_handler.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  robot_scheduler_internal_model_dto.DeviceCreateRequest:
    properties:
      company:
        allOf:
        - $ref: '#/definitions/robot_scheduler_internal_model_entity.CompanyType'
        description: 设备厂商
      extraInfo:
        description: 扩展信息
        type: string
      ip:
        description: 设备IP
        type: string
      password:
        description: 登录密码
        type: string
      port:
        description: 设备端口
        type: integer
      type:
        allOf:
        - $ref: '#/definitions/robot_scheduler_internal_model_entity.DeviceType'
        description: 设备类型
      userName:
        description: 登录用户名
        type: string
    required:
    - company
    - port
    - type
    type: object
  robot_scheduler_internal_model_dto.DeviceUpdateRequest:
    properties:
      company:
        allOf:
        - $ref: '#/definitions/robot_scheduler_internal_model_entity.CompanyType'
        description: 设备厂商
      extraInfo:
        description: 扩展信息
        type: string
      ip:
        description: 设备IP
        type: string
      password:
        description: 登录密码
        type: string
      port:
        description: 设备端口
        type: integer
      status:
        allOf:
        - $ref: '#/definitions/robot_scheduler_internal_model_entity.DeviceStatus'
        description: 设备状态
      type:
        allOf:
        - $ref: '#/definitions/robot_scheduler_internal_model_entity.DeviceType'
        description: 设备类型
      userName:
        description: 登录用户名
        type: string
    type: object
...
  robot_scheduler_internal_model_entity.CompanyType:
    enum:
    - cyborg
    type: string
    x-enum-comments:
      CompanyCyborg: 赛博格
    x-enum-descriptions:
    - 赛博格
    x-enum-varnames:
    - CompanyCyborg
  robot_scheduler_internal_model_entity.DeviceStatus:
    enum:
    - offline
    - online
    - busy
    - error
    type: string
    x-enum-comments:
      DeviceStatusBusy: 忙碌
      DeviceStatusError: 错误
      DeviceStatusOffline: 离线
      DeviceStatusOnline: 在线
    x-enum-descriptions:
    - 离线
    - 在线
    - 忙碌
    - 错误
    x-enum-varnames:
    - DeviceStatusOffline
    - DeviceStatusOnline
    - DeviceStatusBusy
    - DeviceStatusError
  robot_scheduler_internal_model_entity.DeviceType:
    enum:
    - robot_wheel
    - robot_biped
    type: string
    x-enum-comments:
      DeviceTypeBipedRobot: 双足机器人
      DeviceTypeWheelRobot: 轮式机器人
    x-enum-descriptions:
    - 轮式机器人
    - 双足机器人
    x-enum-varnames:
    - DeviceTypeWheelRobot
    - DeviceTypeBipedRobot
...
host: localhost:8080
info:
  contact: {}
  description: 机器人调度系统后台管理接口
  termsOfService: http://swagger.io/terms/
  title: 机器人调度系统 API
  version: "1.0"
paths:
...
  /devices:
    get:
      consumes:
      - application/json
      description: 查询所有设备
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
      security:
      - BearerAuth: []
      summary: 查询设备列表
      tags:
      - 设备管理
    post:
      consumes:
      - application/json
      description: 创建新设备
      parameters:
      - description: 设备信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/robot_scheduler_internal_model_dto.DeviceCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
      security:
      - BearerAuth: []
      summary: 创建设备
      tags:
      - 设备管理
  /devices/{id}:
    delete:
      consumes:
      - application/json
      description: 删除设备（软删除）
      parameters:
      - description: 设备ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
      security:
      - BearerAuth: []
      summary: 删除设备
      tags:
      - 设备管理
    get:
      consumes:
      - application/json
      description: 根据ID获取设备信息
      parameters:
      - description: 设备ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "404":
          description: 设备不存在
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
      security:
      - BearerAuth: []
      summary: 获取设备
      tags:
      - 设备管理
    put:
      consumes:
      - application/json
      description: 更新设备信息
      parameters:
      - description: 设备ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/robot_scheduler_internal_model_dto.DeviceUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "404":
          description: 设备不存在
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
      security:
      - BearerAuth: []
      summary: 更新设备
      tags:
      - 设备管理
...
  /operations:
    get:
      consumes:
      - application/json
      description: 查询所有操作记录
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
      security:
      - BearerAuth: []
      summary: 查询操作记录列表
      tags:
      - 操作记录
  /operations/{id}:
    get:
      consumes:
      - application/json
      description: 根据ID获取操作记录信息
      parameters:
      - description: 操作记录ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "404":
          description: 操作记录不存在
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/internal_api_handler.Response'
      security:
      - BearerAuth: []
      summary: 获取操作记录
      tags:
      - 操作记录
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"